the_newshub/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
crispy-bootstrap5==2025.6
Django==5.2.7
django-ckeditor-5==0.2.18
django-crispy-forms==2.4
django-js-asset==3.1.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
iniconfig==2.3.0
mysqlclient==2.2.7
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
python-dotenv==1.1.1
requests==2.32.5
requests-oauthlib==2.0.0
sqlparse==0.5.3
tweepy==4.16.0
tzdata==2025.2
urllib3==2.5.0
